Updated for new version of Instapaper Extras: Keyboard Shortcuts
The new version allows you to specify a binding to open the Instapaper website.
In addition, security of extras in the main extension has been enhanced.
In the options page, the extension will list all extras, even those which have not made requests.
That way, if a rogue extension is trying to mess with this extension, it will appear in the options panel.
The user would be able to track down the culprit with the extension ID.
In addition, the number of requests (including bogus ones) is counted and tallied up.
For instance, if a buggy extra is spamming requests, it can easily be tracked down.
This is also good for debugging purposes to check if an extra is communicating with the main extension.